Pulisce codice ed elimina alcuni file inutili aggiunti nel .gitignore
