+++ src/Entity/Core/Address.php
@@ -14,19 +14,19 @@ use Doctrine\ORM\Mapping\Embeddable;
 class Address implements \Stringable
 {
     #[Column(name: 'city', type: 'string', length: 256, nullable: true)]
-    protected $city;
+    protected string $city;
 
     #[Column(name: 'province', type: 'string', length: 256, nullable: true)]
-    protected $province;
+    protected string $province;
 
     #[Column(name: 'postal_code', type: 'string', length: 5, nullable: true)]
-    protected $postalCode;
+    protected string $postalCode;
 
     #[Column(name: 'address_line1', type: 'string', length: 256, nullable: true)]
-    protected $addressLine1;
+    protected string $addressLine1;
 
     #[Column(name: 'address_line2', type: 'string', length: 256, nullable: true)]
-    protected $addressLine2;
+    protected string $addressLine2;
 
     public function __construct(
         $city,
