+++ src/Entity/Leader.php
@@ -2,14 +2,14 @@
 
 namespace App\Entity;
 
-use App\Entity\Core\Traits\SoftDeleteableEntity;
-use App\Entity\Core\Traits\TimestampableEntity;
 use App\Entity\Enum\Role;
-use App\Repository\LeaderRepository;
+use Doctrine\ORM\Mapping\Id;
 use Doctrine\ORM\Mapping as ORM;
 use Doctrine\ORM\Mapping\Column;
-use Doctrine\ORM\Mapping\Id;
 use Symfony\Component\Uid\UuidV4;
+use App\Repository\LeaderRepository;
+use App\Entity\Core\Traits\TimestampableEntity;
+use App\Entity\Core\Traits\SoftDeleteableEntity;
 
 #[ORM\Entity(repositoryClass: LeaderRepository::class)]
 class Leader extends Person
@@ -21,7 +21,7 @@ class Leader extends Person
     #[Column(type: 'uuid', unique: true)]
     private UuidV4 $id;
 
-    #[ORM\ManyToOne(inversedBy: 'leaders')]
+    #[ORM\OneToMany(mappedBy: 'leaders')]
     #[ORM\JoinColumn(nullable: false)]
     private ?RoleInUnit $roleInUnit = null;
