+++ src/Entity/OrganizationInterface.php
@@ -4,11 +4,8 @@ declare(strict_types=1);
 
 namespace App\Entity;
 
-use App\Entity\Core\AddressInterface;
-use App\Entity\Core\EmailInterface;
-use App\Entity\Core\PhoneNumber;
-use App\Entity\Core\VisibleInterface;
 use Symfony\Component\Uid\UuidV4;
+use App\Entity\Core\VisibleInterface;
 
 /**
  * Interface OrganizationInterface.
@@ -19,28 +16,11 @@ interface OrganizationInterface extends VisibleInterface
 
     public function getName(): string;
 
-    public function getBusinessName(): string;
-
     public function getDescription(): string;
-
-    public function getPhoneNumber(): ?PhoneNumber;
-
-    public function getEmail(): ?EmailInterface;
-
-    public function getWebsite(): ?string;
-
-    public function getAddress(): AddressInterface;
-
-    public function getSlug(): string;
 
     public function update(
         string $name,
-        string $businessName,
         string $description,
-        ?PhoneNumber $phoneNumber,
-        ?EmailInterface $email,
-        ?string $website,
-        AddressInterface $address,
     ): self;
 
     /**
