+++ src/Entity/RoleInUnit.php
@@ -3,13 +3,13 @@
 namespace App\Entity;
 
 use App\Entity\Enum\Role;
-use App\Repository\RoleInUnitRepository;
-use Doctrine\Common\Collections\ArrayCollection;
-use Doctrine\Common\Collections\Collection;
+use Doctrine\ORM\Mapping\Id;
 use Doctrine\ORM\Mapping as ORM;
 use Doctrine\ORM\Mapping\Column;
-use Doctrine\ORM\Mapping\Id;
 use Symfony\Component\Uid\UuidV4;
+use App\Repository\RoleInUnitRepository;
+use Doctrine\Common\Collections\Collection;
+use Doctrine\Common\Collections\ArrayCollection;
 
 #[ORM\Entity(repositoryClass: RoleInUnitRepository::class)]
 class RoleInUnit
@@ -28,7 +28,7 @@ class RoleInUnit
     /**
      * @var Collection<int, Leader>
      */
-    #[ORM\OneToMany(targetEntity: Leader::class, mappedBy: 'roleInUnit')]
+    #[ORM\ManyToOne(targetEntity: Leader::class, inversedBy: 'roleInUnit')]
     private Collection $leaders;
 
     public function __construct()
