+++ src/Entity/Unit.php
@@ -2,17 +2,17 @@
 
 namespace App\Entity;
 
-use App\Entity\Core\Traits\CollectionsTrait;
-use App\Entity\Core\Traits\SoftDeleteableEntity;
-use App\Entity\Core\Traits\TimestampableEntity;
+use Doctrine\ORM\Mapping\Id;
 use App\Entity\Enum\UnitType;
 use Doctrine\ORM\Mapping\Column;
 use Doctrine\ORM\Mapping\Entity;
-use Doctrine\ORM\Mapping\Id;
-use Doctrine\ORM\Mapping\JoinColumn;
-use Doctrine\ORM\Mapping\ManyToOne;
-use Gedmo\Mapping\Annotation as Gedmo;
 use Symfony\Component\Uid\UuidV4;
+use Doctrine\ORM\Mapping\ManyToOne;
+use Doctrine\ORM\Mapping\JoinColumn;
+use Gedmo\Mapping\Annotation as Gedmo;
+use App\Entity\Core\Traits\CollectionsTrait;
+use App\Entity\Core\Traits\TimestampableEntity;
+use App\Entity\Core\Traits\SoftDeleteableEntity;
 
 #[Entity]
 class Unit implements UnitInterface
@@ -34,7 +34,7 @@ class Unit implements UnitInterface
     #[Column(length: 128)]
     private string $slug;
 
-    #[ManyToOne(targetEntity: 'App\Entity\ScoutGroup', inversedBy: 'units')]
+    #[ManyToOne(targetEntity: 'App\Entity\Group', inversedBy: 'units')]
     #[JoinColumn(name: 'organization_id', referencedColumnName: 'id')]
     private OrganizationInterface $group;
