+++ src/Repository/RoleInUnitRepository.php
@@ -3,8 +3,8 @@
 namespace App\Repository;
 
 use App\Entity\RoleInUnit;
-use Doctrine\Bundle\DoctrineBundle\Repository\ServiceEntityRepository;
 use Doctrine\Persistence\ManagerRegistry;
+use Doctrine\Bundle\DoctrineBundle\Repository\ServiceEntityRepository;
 
 /**
  * @extends ServiceEntityRepository<RoleInUnit>
@@ -15,29 +15,4 @@ class RoleInUnitRepository extends ServiceEntityRepository
     {
         parent::__construct($registry, RoleInUnit::class);
     }
-
-    //    /**
-    //     * @return RoleInUnit[] Returns an array of RoleInUnit objects
-    //     */
-    //    public function findByExampleField($value): array
-    //    {
-    //        return $this->createQueryBuilder('r')
-    //            ->andWhere('r.exampleField = :val')
-    //            ->setParameter('val', $value)
-    //            ->orderBy('r.id', 'ASC')
-    //            ->setMaxResults(10)
-    //            ->getQuery()
-    //            ->getResult()
-    //        ;
-    //    }
-
-    //    public function findOneBySomeField($value): ?RoleInUnit
-    //    {
-    //        return $this->createQueryBuilder('r')
-    //            ->andWhere('r.exampleField = :val')
-    //            ->setParameter('val', $value)
-    //            ->getQuery()
-    //            ->getOneOrNullResult()
-    //        ;
-    //    }
 }
